backend/pyproject.toml
[project]
name = "super-bryn"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "groq>=1.0.0",
    "livekit-agents[groq,mcp,silero,tavus,turn-detector]~=1.3",
    "livekit-api>=1.0.0",
    "livekit-plugins-groq>=1.4.1",
    "livekit-plugins-noise-cancellation~=0.2",
    "onnxruntime<1.19.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
]
